replace all int/uint with isize/usize

+++ src/geom.rs
@@ -48,8 +48,8 @@ mod test {
 
     #[test]
     fn test_add() {
-        let p1 = Point::<int>::new(2, 3);
-        let p2 = Point::<int>::new(1, 5);
+        let p1 = Point::<isize>::new(2, 3);
+        let p2 = Point::<isize>::new(1, 5);
         let p3 = p1 + p2;
         assert_eq!((p3.x, p3.y), (3, 8));
     }
@@ -57,8 +57,8 @@ mod test {
 
     #[test]
     fn test_sub() {
-        let p1 = Point::<int>::new(2, 3);
-        let p2 = Point::<int>::new(1, 5);
+        let p1 = Point::<isize>::new(2, 3);
+        let p2 = Point::<isize>::new(1, 5);
         let p3 = p1 - p2;
         assert_eq!((p3.x, p3.y), (1, -2));
     }
@@ -66,7 +66,7 @@ mod test {
 
     #[test]
     fn test_scale() {
-        let p = Point::<int>::new(2, 3).scale(2);
+        let p = Point::<isize>::new(2, 3).scale(2);
         assert_eq!((p.x, p.y), (4, 6));
     }
 }
